mcp-proxmox/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
proxmoxer>=2.0.1
requests>=2.26.0
python-dotenv>=0.19.0
pydantic>=1.8.2
